pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parts-mcp"
version = "0.1.3"
description = "Model Context Protocol (MCP) server for electronic parts sourcing with KiCad integration"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Parts MCP Contributors" }
]
maintainers = [
    { name = "Parts MCP Contributors" }
]
keywords = [
    "electronics",
    "parts",
    "components",
    "sourcing",
    "kicad",
    "mcp",
    "model-context-protocol",
    "bom",
    "pcb"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "fastmcp>=2.13.0,<3.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "diskcache>=5.0.0",
    "pandas>=2.0.0",
    "defusedxml>=0.7.0",
    "pyyaml>=6.0",
]

[project.urls]
"Homepage" = "https://github.com/SourceParts/parts-mcp"
"Bug Tracker" = "https://github.com/SourceParts/parts-mcp/issues"
"Documentation" = "https://source.parts/docs/api"
"Source" = "https://github.com/SourceParts/parts-mcp"

[project.optional-dependencies]
hosted = [
    "uvicorn>=0.30.0",
    "starlette>=0.37.0",
    "py-key-value-aio[redis]>=0.1.0",
    "cryptography>=42.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
parts-mcp = "parts_mcp.server:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by ruff format
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["parts_mcp"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "mcp.*",
    "diskcache.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]